Reduce memory use (O2O3O4)
