+++ src/symfc/solvers/solver_O3O4.py
@@ -10,12 +10,15 @@ import numpy as np
 
 from symfc.basis_sets import FCBasisSetO3, FCBasisSetO4
 from symfc.eig_solvers.matrix import block_matrix_sandwich
-from symfc.solvers.solver_O2O3 import reshape_nNN333_nx_to_N3N3_n3nx, set_disps_N3N3
-from symfc.solvers.solver_O2O3O4 import (
-    reshape_nNNN3333_nx_to_N3N3N3_n3nx,
+from symfc.utils.solver_funcs import get_batch_slice, solve_linear_equation
+from symfc.utils.solver_utils_O3 import (
+    reshape_compr_mat_O3,
+    set_disps_N3N3,
+)
+from symfc.utils.solver_utils_O4 import (
+    reshape_compr_mat_O4,
     set_disps_N3N3N3,
 )
-from symfc.utils.solver_funcs import get_batch_slice, solve_linear_equation
 
 try:
     from symfc.utils.matrix import dot_product_sparse
@@ -203,8 +206,6 @@ def prepare_normal_equation_O3O4(
     """
     N3 = disps.shape[1]
     N = N3 // 3
-    NN = N**2
-    NNN = N**3
 
     n_compr_fc3 = compact_compress_mat_fc3.shape[1]
     n_compr_fc4 = compact_compress_mat_fc4.shape[1]
@@ -229,27 +230,13 @@ def prepare_normal_equation_O3O4(
         if verbose:
             print("-----", flush=True)
             print("Solver_atoms:", begin_i + 1, "--", end_i, "/", N, flush=True)
-        n_atom_batch = end_i - begin_i
 
         t1 = time.time()
-        decompr_idx = (
-            atomic_decompr_idx_fc3[begin_i * NN : end_i * NN, None] * 27
-            + np.arange(27)[None, :]
-        ).reshape(-1)
-        compr_mat_fc3 = reshape_nNN333_nx_to_N3N3_n3nx(
-            compact_compress_mat_fc3[decompr_idx],
-            N,
-            n_atom_batch,
+        compr_mat_fc3 = reshape_compr_mat_O3(
+            compact_compress_mat_fc3, atomic_decompr_idx_fc3, N, begin_i, end_i
         )
-
-        decompr_idx = (
-            atomic_decompr_idx_fc4[begin_i * NNN : end_i * NNN, None] * 81
-            + np.arange(81)[None, :]
-        ).reshape(-1)
-        compr_mat_fc4 = reshape_nNNN3333_nx_to_N3N3N3_n3nx(
-            compact_compress_mat_fc4[decompr_idx],
-            N,
-            n_atom_batch,
+        compr_mat_fc4 = reshape_compr_mat_O4(
+            compact_compress_mat_fc4, atomic_decompr_idx_fc4, N, begin_i, end_i
         )
         t2 = time.time()
         if verbose:
